retire pidgin
