+++ services/apis/peregrine/peregrineProps.js
@@ -8,5 +8,7 @@ module.exports = {
   endpoints: {
     version: '/api/search/_version',
     query: '/api/v0/submission/graphql',
+    coreMetadataPath: '/api/search/coremetadata',
+    coreMetadataLegacyPath: '/coremetadata',
   },
 };
