--- services/apis/pidgin/pidginProps.js
@@ -1,9 +1,0 @@
-/**
- * pidgin Properties
- */
-module.exports = {
-  endpoints: {
-    coreMetadataPath: '/api/search/coremetadata',
-    coreMetadataLegacyPath: '/coremetadata',
-  },
-};
