--- services/apis/pidgin/pidginSequences.js
@@ -1,9 +1,0 @@
-// const pidginQuestions = require('./pidginQuestions.js');
-// const pidginTasks = require('./pidginTasks.js');
-
-/**
- * pidgin sequences
- */
-module.exports = {
-  // Sequences are for an service to combine multiple tasks and questions
-};
