+++ suites/core-metadata/core-metadata.feature
@@ -1,4 +1,4 @@
-@Pidgintests
+@CoreMetadataTests
 Feature : GetCoremetadata 
 
     @JsonFormat
